[connectionagent] Use connman enums. JB#61909
Removed storing the service state to the config file. As transient
state it was strange there. Was added with commit 9eeeae113 where
it actually read the value back, but these days this was only
writing.

The old code was also slightly mixing up network manager and network
service states on connmanAvailabilityChanged(), but then not actually
using the stored network service state anywhere. Removed that too.